+++ UnitTests/BraintreeApplePayTests/BraintreeApplePayTests-Bridging-Header.h
@@ -1,1 +1,2 @@
 #import <BraintreeApplePay/BTApplePayClient_Internal.h>
+#import <BraintreeApplePay/BTApplePayCardNonce.h>