+++ Sources/BraintreeApplePay/Public/BraintreeApplePay/BTApplePayCardNonce.h
@@ -19,7 +19,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  Used to initialize a `BTApplePayCardNonce` with parameters.
  */
-- (nullable instancetype)initWithNonce:(NSString *)nonce type:(NSString *)type json:(BTJSON *)json;
+- (nullable instancetype)initWithJSON:(BTJSON *)json;
 
 @end