+++ SampleApps/SPMTest/SPMTest/Frameworks/CardinalMobile.framework/Headers/CardinalMobile.h
@@ -1,0 +1,20 @@
+//
+//  CardinalMobile.h
+//  CardinalMobile
+//
+//  Created by Sudeep Tuladhar on 10/23/18.
+//  Copyright © 2018 Cardinal Commerce. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for CardinalMobile.
+FOUNDATION_EXPORT double CardinalMobileVersionNumber;
+
+//! Project version string for CardinalMobile.
+FOUNDATION_EXPORT const unsigned char CardinalMobileVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <CardinalMobile/PublicHeader.h>
+
+#import <CardinalMobile/CardinalCMSDK.h>
+#import <CardinalMobile/CardinalEMVCoSDK.h>