+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased (5.0.0-beta2)
 * Add SPM support for `BraintreeDataCollector`
 * Bump Kount to v4.0.4.3 pre-release (provides an xcframework for SPM)
+* Breaking Changes
+  * Remove `type` and `nonce` params on `BTApplePayCardNone` initializer
 
 ## 5.0.0-beta1 (2020-12-01)
 * Add support for Swift Package Manager (resolves #462)